refactor: move form submit data to redux
