+++ 01-components/src/pages/FormPage.tsx
@@ -1,14 +1,18 @@
 import RequestForm from 'components/Form';
 import RequestCard from 'components/MovieReqCard';
-import { UserContext } from 'components/UserContext';
+import { useAppDispatch, useAppSelector } from 'app/hooks';
 import React from 'react';
 import { Link } from 'react-router-dom';
 import Footer from '../components/Footer';
 import Header from '../components/Header';
+import { addRequestData } from 'app/requestsSlice';
+import { MovieReqData } from 'utils/TMDBinterfaces';
 
 export default function FormPage() {
-  const { movieReqData, addRequest } = React.useContext(UserContext);
-  const cards = movieReqData.map((item) => <RequestCard key={item.num} {...item} />);
+  const requests = useAppSelector((state) => state.requests.requests);
+  const dispatch = useAppDispatch();
+  const requestFormHandler = (data: MovieReqData) => dispatch(addRequestData(data));
+  const cards = requests.map((item) => <RequestCard key={item.num} {...item} />);
 
   return (
     <div className="root">
@@ -16,7 +20,7 @@ export default function FormPage() {
       <div className="used-cameras-container">
         <div className="used-cameras-left">
           <h2 className="used-cameras-heading">Add your request for a movie</h2>
-          <RequestForm movieReq={addRequest} />
+          <RequestForm movieReq={requestFormHandler} />
           <Link to="/">
             <button className="button_big">Return to main</button>
           </Link>
