+++ 01-components/src/utils/TMDBinterfaces.ts
@@ -85,3 +85,21 @@ export interface MovieResponse {
   total_results: number;
   total_pages: number;
 }
+
+export interface MovieReqData {
+  num?: number;
+  date: string;
+  description: string;
+  files: FileList;
+  genres: {
+    action: boolean;
+    comedy: boolean;
+    drama: boolean;
+    fantasy: boolean;
+    other: boolean;
+  };
+  title: string;
+  score: string;
+  type: string;
+  viewed: boolean;
+}
